fix: useRef for setTimeout
