+++ packages/ui/src/assets/images/onboarding/appUpdate/index.ts
@@ -1,1 +1,2 @@
 export * from './AppUpdateIcon';
+export * from './LoadingIcon';
