+++ packages/ui/src/components/molecules/CopyContainer.tsx
@@ -1,8 +1,8 @@
-import React, { FC, useState } from 'react';
+import React, { FC } from 'react';
 import styled from 'styled-components';
 import { utils, UtilsProps } from '../utils';
-import { Button, Image, Typography } from '../atoms';
-import { checkIcon, copyIcon } from '../../assets';
+import { Typography } from '../atoms';
+import { Clipboard } from './Clipboard';
 
 interface CopyContainerProps extends UtilsProps {
   link: string;
@@ -20,29 +20,11 @@ const MaskStyle = styled.div<Omit<CopyContainerProps, 'link'>>`
   padding: 12px 16px;
   ${utils}
 `;
-export const CopyContainer: FC<CopyContainerProps> = ({ link, ...props }) => {
-  const [copied, setCopied] = useState(false);
-  const handleClick = async () => {
-    setCopied(true);
-    const timeout = setTimeout(() => {
-      setCopied(false);
-      if (timeout) clearTimeout(timeout);
-    }, 2000);
-    await navigator.clipboard.writeText(link);
-  };
-
-  return (
-    <MaskStyle {...props}>
-      <Typography color="muted" variant="p">
-        {link}
-      </Typography>
-      {!copied ? (
-        <Button onClick={handleClick} variant="none" display="flex">
-          <Image src={copyIcon} height="20" alt="Copy" />
-        </Button>
-      ) : (
-        <Image src={checkIcon} alt="copied" />
-      )}
-    </MaskStyle>
-  );
-};
+export const CopyContainer: FC<CopyContainerProps> = ({ link, ...props }) => (
+  <MaskStyle {...props}>
+    <Typography color="muted" variant="p">
+      {link}
+    </Typography>
+    <Clipboard content={link} size="md" />
+  </MaskStyle>
+);
